+++ src/components/Blender/CheckboxKeyword/CheckboxKeyword.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import _ from 'lodash';
-import * as formActions from '../../../actions/formActions';
+import * as formActions from '../formActions';
 
 /**
  * Component for one theme's checkbox
@@ -61,8 +61,6 @@ const mapDispatchToProps = dispatch => ({
   },
 });
 
-const CheckboxKeyword = connect(mapStateToProps, mapDispatchToProps)(
-  CheckboxKeywordComponent,
-);
+const CheckboxKeyword = connect(mapStateToProps, mapDispatchToProps)(CheckboxKeywordComponent);
 
 export default CheckboxKeyword;
